requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
-e git+https://github.com/mi-24v/Flask-FirebaseAuth@e3e29f6a803b115c33ed224574ffd5929f254d35#egg=Flask_FirebaseAuth
Flask-JSON==0.3.4
Flask-Login==0.4.1
gast==0.2.2
google-api-core==1.14.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
namegenerator==1.0.6
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2